[TASK] Deprecate `__toString()`
Part of #998
