+++ src/CSSList/AtRuleBlockList.php
@@ -42,6 +42,9 @@ class AtRuleBlockList extends CSSBlockList implements AtRule
         return $this->arguments;
     }
 
+    /**
+     * @deprecated in V8.8.0, will be removed in V9.0.0. Use `render` instead.
+     */
     public function __toString(): string
     {
         return $this->render(new OutputFormat());
