+++ src/CSSList/CSSList.php
@@ -402,6 +402,9 @@ abstract class CSSList implements Renderable, Commentable
         }
     }
 
+    /**
+     * @deprecated in V8.8.0, will be removed in V9.0.0. Use `render` instead.
+     */
     public function __toString(): string
     {
         return $this->render(new OutputFormat());
