+++ src/CSSList/KeyFrame.php
@@ -51,6 +51,9 @@ class KeyFrame extends CSSList implements AtRule
         return $this->animationName;
     }
 
+    /**
+     * @deprecated in V8.8.0, will be removed in V9.0.0. Use `render` instead.
+     */
     public function __toString(): string
     {
         return $this->render(new OutputFormat());
