+++ src/Comment/Comment.php
@@ -50,6 +50,9 @@ class Comment implements Renderable
         $this->commentText = $commentText;
     }
 
+    /**
+     * @deprecated in V8.8.0, will be removed in V9.0.0. Use `render` instead.
+     */
     public function __toString(): string
     {
         return $this->render(new OutputFormat());
