+++ src/Property/CSSNamespace.php
@@ -54,6 +54,9 @@ class CSSNamespace implements AtRule
         return $this->lineNumber;
     }
 
+    /**
+     * @deprecated in V8.8.0, will be removed in V9.0.0. Use `render` instead.
+     */
     public function __toString(): string
     {
         return $this->render(new OutputFormat());
