+++ src/Property/Charset.php
@@ -71,6 +71,9 @@ class Charset implements AtRule
         return $this->charset->getString();
     }
 
+    /**
+     * @deprecated in V8.8.0, will be removed in V9.0.0. Use `render` instead.
+     */
     public function __toString(): string
     {
         return $this->render(new OutputFormat());
