+++ src/Property/Import.php
@@ -72,6 +72,9 @@ class Import implements AtRule
         return $this->location;
     }
 
+    /**
+     * @deprecated in V8.8.0, will be removed in V9.0.0. Use `render` instead.
+     */
     public function __toString(): string
     {
         return $this->render(new OutputFormat());
