+++ src/Property/Selector.php
@@ -113,6 +113,9 @@ class Selector
         $this->specificity = null;
     }
 
+    /**
+     * @deprecated in V8.8.0, will be removed in V9.0.0. Use `render` instead.
+     */
     public function __toString(): string
     {
         return $this->getSelector();
