+++ src/Renderable.php
@@ -6,6 +6,9 @@ namespace Sabberworm\CSS;
 
 interface Renderable
 {
+    /**
+     * @deprecated in V8.8.0, will be removed in V9.0.0. Use `render` instead.
+     */
     public function __toString(): string;
 
     public function render(OutputFormat $outputFormat): string;
