+++ src/Rule/Rule.php
@@ -260,6 +260,9 @@ class Rule implements Renderable, Commentable
         return $this->isImportant;
     }
 
+    /**
+     * @deprecated in V8.8.0, will be removed in V9.0.0. Use `render` instead.
+     */
     public function __toString(): string
     {
         return $this->render(new OutputFormat());
