+++ src/RuleSet/AtRuleSet.php
@@ -53,6 +53,9 @@ class AtRuleSet extends RuleSet implements AtRule
         return $this->arguments;
     }
 
+    /**
+     * @deprecated in V8.8.0, will be removed in V9.0.0. Use `render` instead.
+     */
     public function __toString(): string
     {
         return $this->render(new OutputFormat());
