+++ src/RuleSet/DeclarationBlock.php
@@ -142,6 +142,8 @@ class DeclarationBlock extends RuleSet
 
     /**
      * @throws OutputException
+     *
+     * @deprecated in V8.8.0, will be removed in V9.0.0. Use `render` instead.
      */
     public function __toString(): string
     {
