+++ src/RuleSet/RuleSet.php
@@ -255,6 +255,9 @@ abstract class RuleSet implements Renderable, Commentable
         }
     }
 
+    /**
+     * @deprecated in V8.8.0, will be removed in V9.0.0. Use `render` instead.
+     */
     public function __toString(): string
     {
         return $this->render(new OutputFormat());
