+++ src/Value/CSSFunction.php
@@ -105,6 +105,9 @@ class CSSFunction extends ValueList
         return $this->aComponents;
     }
 
+    /**
+     * @deprecated in V8.8.0, will be removed in V9.0.0. Use `render` instead.
+     */
     public function __toString(): string
     {
         return $this->render(new OutputFormat());
