+++ src/Value/CSSString.php
@@ -90,6 +90,9 @@ class CSSString extends PrimitiveValue
         return $this->string;
     }
 
+    /**
+     * @deprecated in V8.8.0, will be removed in V9.0.0. Use `render` instead.
+     */
     public function __toString(): string
     {
         return $this->render(new OutputFormat());
