+++ src/Value/Color.php
@@ -224,6 +224,9 @@ class Color extends CSSFunction
         return $this->getName();
     }
 
+    /**
+     * @deprecated in V8.8.0, will be removed in V9.0.0. Use `render` instead.
+     */
     public function __toString(): string
     {
         return $this->render(new OutputFormat());
