+++ src/Value/LineName.php
@@ -49,6 +49,9 @@ class LineName extends ValueList
         return new LineName($aNames, $parserState->currentLine());
     }
 
+    /**
+     * @deprecated in V8.8.0, will be removed in V9.0.0. Use `render` instead.
+     */
     public function __toString(): string
     {
         return $this->render(new OutputFormat());
