+++ src/Value/Size.php
@@ -210,6 +210,9 @@ class Size extends PrimitiveValue
         return false;
     }
 
+    /**
+     * @deprecated in V8.8.0, will be removed in V9.0.0. Use `render` instead.
+     */
     public function __toString(): string
     {
         return $this->render(new OutputFormat());
