+++ src/Value/URL.php
@@ -76,6 +76,9 @@ class URL extends PrimitiveValue
         return $this->oURL;
     }
 
+    /**
+     * @deprecated in V8.8.0, will be removed in V9.0.0. Use `render` instead.
+     */
     public function __toString(): string
     {
         return $this->render(new OutputFormat());
