+++ src/Value/ValueList.php
@@ -83,6 +83,9 @@ abstract class ValueList extends Value
         $this->sSeparator = $sSeparator;
     }
 
+    /**
+     * @deprecated in V8.8.0, will be removed in V9.0.0. Use `render` instead.
+     */
     public function __toString(): string
     {
         return $this->render(new OutputFormat());
